V4.0 of waa2wls-protocol.txt - define kid better
'Clarify' the syntax of the 'kid' field in response messages to make it
clear that it must be numeric, between 1 and 8 characters long, and
without leading '0' characters. This matches current reality, and
makes it easier to address path traversal vulnerabilities in agents
that use the kit to construct the path to a key file.