Fix a lurking typo implying that response fields are ':'-seperated
The description of the 'sig' field in the response message contained a
reference to response message fields being seperated by ':'. Correted
this to '!'. Once upon a time, the fields in the response message were
seperated by ':'and this typo was presumably inherited from then and
has remained undetected for over 10 years.